separate CallTimeGenerator test into its own test class
